requirements.txt
Flask
paho-mqtt
pymongo
